fix the montant_disponible issue

+++ database/migrations/0000_04_18_115917_create_centres_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
            $table->string('adresse_centre',200)->nullable();
            $table->integer('seuil_centre')->nullable();
            $table->enum('type_centre',['Aviation','Marine'])->nullable();
-            $table->integer('montant_disponible')->nullable();
+            $table->decimal('montant_disponible')->nullable();
         });
     }
