+++ database/migrations/2025_04_19_214515_create_dras_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreign('id_centre')->references('id_centre')->on('centres')->onDelete('cascade');
             $table->date('date_creation');
             $table->enum('etat', ['actif', 'cloture','refuse','accepte','rembourse'])->default('actif');
-            $table->integer('total_dra');
+            $table->decimal('total_dra');
             $table->timestamp('created_at')->nullable();
         });
     }
